+++ src/Service/SchemaCompilerInterface.php
@@ -8,7 +8,6 @@ use Cycle\Database\DatabaseManager;
 use Cycle\ORM\Exception\ConfigException;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
-use Sirix\Cycle\Enum\SchemaCompileMode;
 
 interface SchemaCompilerInterface
 {
@@ -28,6 +27,41 @@ interface SchemaCompilerInterface
         array $entities,
         array $manualMappingSchemaDefinitions,
         array $additionalGenerators = [],
-        SchemaCompileMode $schemaCompileMode = SchemaCompileMode::Runtime,
+    ): array;
+
+    /**
+     * @param array<string>        $entities
+     * @param array<string, mixed> $manualMappingSchemaDefinitions
+     * @param array<int, mixed>    $additionalGenerators
+     *
+     * @return array<string, mixed>
+     *
+     * @throws ConfigException
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     */
+    public function sync(
+        DatabaseManager $databaseManager,
+        array $entities,
+        array $manualMappingSchemaDefinitions,
+        array $additionalGenerators = [],
+    ): array;
+
+    /**
+     * @param array<string>        $entities
+     * @param array<string, mixed> $manualMappingSchemaDefinitions
+     * @param array<int, mixed>    $additionalGenerators
+     *
+     * @return array<string, mixed>
+     *
+     * @throws ConfigException
+     * @throws ContainerExceptionInterface
+     * @throws NotFoundExceptionInterface
+     */
+    public function generateMigrations(
+        DatabaseManager $databaseManager,
+        array $entities,
+        array $manualMappingSchemaDefinitions,
+        array $additionalGenerators = [],
     ): array;
 }